refact(TierSelector): migrate to vanilla extract css
